+++ DspLib/dspSlowThresholderParms.h
@@ -34,16 +34,16 @@ public:
    // Members.
 
    // Fuzzy alpha filter sampling period and time constant, seconds.
-   double mAlphaFilterTs;
-   double mAlphaFilterTc;
+   float mAlphaFilterTs;
+   float mAlphaFilterTc;
 
    // Fuzzy to crisp threshold.
-   double mFuzzyToCrispThreshLo;
-   double mFuzzyToCrispThreshHi;
+   float mFuzzyToCrispThreshLo;
+   float mFuzzyToCrispThreshHi;
 
    // Signal thresholds.
-   double mValueThreshLo;
-   double mValueThreshHi;
+   float mValueThreshLo;
+   float mValueThreshHi;
 
    //***************************************************************************
    //***************************************************************************
